"Importing" from RCS:  .init-tasks.sh, rev. 1.12
Since SVN doesn't do direct import of RCS, I've appended the logs &
diffs of every revision from 1.1 to 1.12 to the end of this file
(search for "# End" to find it).

Each revision has its log printed, followed by a line of "===...===",
followed by an "rcsdiff" between the logged revision and the previous
one.

NOTE:  These diffs will be removed in the next commit!


git-svn-id: file:///home/candide/src/0_Local_SVN_Repos/sh.scripts/trunk@1952 92e5ae4a-8afd-0310-8b33-eaebc6dd6768